refactor product sell form to use hard sql. Product type populates a drop down on sell form. Product is posting to DB
