+++ .gitignore
@@ -56,4 +56,4 @@ scratchNotes/
 Pipfile*
 migrations
 db.sqlite3-journal
-# sqlite3.sql
+sqlite3.sql
