+++ website/views/__init__.py
@@ -1,1 +1,2 @@
-from .views import index, register, login_user, user_logout, sell_product, list_products
+from .views import index, register, login_user, user_logout, list_products
+from .product_category_view import sell_product
